Moved unit tests to unit test directory
