+++ tests/unit/settlements.js
@@ -36,37 +36,37 @@ const Test = require('tape');
 const Hapi = require('hapi');
 const HapiOpenAPI = require('hapi-openapi');
 const Path = require('path');
-const Mockgen = require('../data/mockgen.js');
+const Mockgen = require('../../data/mockgen.js');
 const responseCodes = [200, 400, 401, 404, 415, 500];
 
 /**
- * Test for /settlementWindows
+ * Test for /settlements
  */
-Test('/settlementWindows', function (t) {
+Test('/settlements', function (t) {
 
     /**
-     * summary: Returns a Settlement Window(s) as per parameter(s).
+     * summary: Returns Settlement(s) as per parameter(s).
      * description:
-     * parameters: participantId, state, fromDateTime, toDateTime
+     * parameters: currency, participantId, settlementWindowId, accountId, state, fromDateTime, toDateTime
      * produces: application/json
      * responses: 200, 400, 401, 404, 415, default
      */
-    t.test('test getSettlementWindowByParams get operation', async function (t) {
+    t.test('test getSettlementsByParams get operation', async function (t) {
 
         const server = new Hapi.Server();
         try {
             await server.register({
                 plugin: HapiOpenAPI,
                 options: {
-                    api: Path.resolve(__dirname, '../config/swagger.json'),
-                    handlers: Path.join(__dirname, '../handlers'),
+                    api: Path.resolve(__dirname, '../../config/swagger.json'),
+                    handlers: Path.join(__dirname, '../../handlers'),
                     outputvalidation: true
                 }
             });
 
             const requests = new Promise((resolve, reject) => {
                 Mockgen().requests({
-                    path: '/settlementWindows',
+                    path: '/settlements',
                     operation: 'get'
                 }, function (error, mock) {
                     return error ? reject(error) : resolve(mock);
@@ -131,6 +131,7 @@ Test('/settlementWindows', function (t) {
             t.fail(e);
             t.end()
         }
+
     });
 
 });
