+++ tests/unit/settlements/{settlementId}/participants/{participantId}.js
@@ -36,38 +36,38 @@ const Test = require('tape');
 const Hapi = require('hapi');
 const HapiOpenAPI = require('hapi-openapi');
 const Path = require('path');
-const Mockgen = require('../data/mockgen.js');
+const Mockgen = require('../../../../../data/mockgen.js');
 const responseCodes = [200, 400, 401, 404, 415, 500];
 
 /**
- * Test for /settlements
+ * Test for /settlements/{settlementId}/participants/{participantId}
  */
-Test('/settlements', function (t) {
+Test('/settlements/{settlementId}/participants/{participantId}', function (t) {
 
     /**
-     * summary: Returns Settlement(s) as per parameter(s).
+     * summary: Acknowledegement of settlement by updating with Settlements Id and Participant Id.
      * description:
-     * parameters: currency, participantId, settlementWindowId, accountId, state, fromDateTime, toDateTime
+     * parameters: settlementId, participantId, settlementParticipantUpdatePayload
      * produces: application/json
      * responses: 200, 400, 401, 404, 415, default
      */
-    t.test('test getSettlementsByParams get operation', async function (t) {
+    t.test('test updateSettlementBySettlementIdParticipantId put operation', async function (t) {
 
         const server = new Hapi.Server();
         try {
             await server.register({
                 plugin: HapiOpenAPI,
                 options: {
-                    api: Path.resolve(__dirname, '../config/swagger.json'),
-                    handlers: Path.join(__dirname, '../handlers'),
+                    api: Path.resolve(__dirname, '../../../../../config/swagger.json'),
+                    handlers: Path.join(__dirname, '../../../../../handlers'),
                     outputvalidation: true
                 }
             });
 
             const requests = new Promise((resolve, reject) => {
                 Mockgen().requests({
-                    path: '/settlements',
-                    operation: 'get'
+                    path: '/settlements/{settlementId}/participants/{participantId}',
+                    operation: 'put'
                 }, function (error, mock) {
                     return error ? reject(error) : resolve(mock);
                 });
@@ -80,7 +80,7 @@ Test('/settlements', function (t) {
             //Get the resolved path from mock request
             //Mock request Path templates({}) are resolved using path parameters
             const options = {
-                method: 'get',
+                method: 'put',
                 url: '/v1' + mock.request.path
             };
             if (mock.request.body) {
