+++ tests/unit/settlements/{settlementId}/participants/{participantId}/accounts/{accountId}.js
@@ -35,7 +35,7 @@ const Test = require('tape');
 const Hapi = require('hapi');
 const HapiOpenAPI = require('hapi-openapi');
 const Path = require('path');
-const Mockgen = require('../../../../../../data/mockgen.js');
+const Mockgen = require('../../../../../../../data/mockgen.js');
 const responseCodes = [200, 400, 401, 404, 415, 500];
 
 /**
@@ -57,8 +57,8 @@ Test('/settlements/{settlementId}/participants/{participantId}/accounts/{account
         await server.register({
             plugin: HapiOpenAPI,
             options: {
-                api: Path.resolve(__dirname, '../../../../../../config/swagger.json'),
-                handlers: Path.join(__dirname, '../../../../../../handlers'),
+                api: Path.resolve(__dirname, '../../../../../../../config/swagger.json'),
+                handlers: Path.join(__dirname, '../../../../../../../handlers'),
                 outputvalidation: true
             }
         });
@@ -146,8 +146,8 @@ Test('/settlements/{settlementId}/participants/{participantId}/accounts/{account
         await server.register({
             plugin: HapiOpenAPI,
             options: {
-                api: Path.resolve(__dirname, '../../../../../../config/swagger.json'),
-                handlers: Path.join(__dirname, '../../../../../../handlers'),
+                api: Path.resolve(__dirname, '../../../../../../../config/swagger.json'),
+                handlers: Path.join(__dirname, '../../../../../../../handlers'),
                 outputvalidation: true
             }
         });
